Moved trend and noise calculations to their own module.
